hrgn... typescript.. :-(
