+++ main.js
@@ -248,94 +248,7 @@ class MyDlink extends utils.Adapter {
      */
     async onReady() {
 
-        //start existing devices:
-        let haveActiveDevices = false;
-        const existingDevices = await this.getDevicesAsync();
-        const configDevicesToAdd = [].concat(this.config.devices);
-        this.log.debug('Got existing devices: ' + JSON.stringify(existingDevices, null, 2));
-        this.log.debug('Got config devices: ' + JSON.stringify(configDevicesToAdd, null, 2));
-        let needUpdateConfig = false;
-        for (const existingDevice of existingDevices) {
-            let found = false;
-            for (const configDevice of this.config.devices) {
-                // @ts-ignore
-                needUpdateConfig = !configDevice.mac;
-                // @ts-ignore
-                if ((configDevice.mac && configDevice.mac === existingDevice.native.mac) ||
-                    // @ts-ignore
-                    (!configDevice.mac && configDevice.ip === existingDevice.native.ip)) {
-                    found = true;
 
-                    //copy all data from config, because now only config is changed from config dialog.
-                    for (const key of Object.keys(configDevice)) {
-                        existingDevice.native[key] = configDevice[key]; //copy all fields.
-                    }
-                    // @ts-ignore
-                    existingDevice.native.pinNotEncrypted = !configDevice.mac;
-
-                    configDevicesToAdd.splice(configDevicesToAdd.indexOf(configDevice), 1);
-                    break; //break on first copy -> will remove additional copies later.
-                }
-            }
-            const device = DeviceInfo.createFromObject(existingDevice);
-            await this.createNewDevice(device); //store new config.
-            if (existingDevice.native.pinNotEncrypted) {
-                needUpdateConfig = true;
-            }
-            if (found) {
-                haveActiveDevices = await this.startDevice(device) || haveActiveDevices;
-                //keep config and client for later reference.
-                this.devices.push(device);
-            } else {
-                this.log.debug('Deleting ' + device.name);
-                await this.deleteDeviceFull(device);
-            }
-        }
-
-        //add non-existing devices from config:
-        for (const configDevice of configDevicesToAdd) {
-            const device = DeviceInfo.createFromTable(configDevice, !configDevice.pinNotEncrypted);
-            this.log.debug('Device ' + device.name + ' in config but not in devices -> create and add.');
-            const oldDevice = this.devices.find(d => d.mac === device.mac);
-            if (oldDevice) {
-                this.log.info('Duplicate entry for ' + device.mac + ' in config. Trying to rectify. Restart will happen. Affected devices: ' + device.name + ' === ' + configDevice.name);
-                needUpdateConfig = true;
-            } else {
-                //make sure objects are created:
-                await this.createNewDevice(device);
-
-                haveActiveDevices = await this.startDevice(device) || haveActiveDevices;
-                //call this here again, to make sure it happens.
-                await this.createNewDevice(device); //store device settings
-                //keep config and client for later reference.
-                this.devices.push(device);
-            }
-        }
-
-        //try to update config:
-        if (needUpdateConfig) {
-            const devices = [];
-            for (const device of this.devices) {
-                const configDevice = {
-                    ip: device.ip,
-                    mac: device.mac,
-                    pin: DeviceInfo.encryptDecrypt(this.secret, device.pin),
-                    pollInterval: device.pollInterval,
-                    enabled: device.enabled,
-                    name: device.name,
-                    model: device.model,
-                    useWebSocket: device.useWebSocket
-                };
-                devices.push(configDevice);
-            }
-            await this.extendForeignObjectAsync('system.adapter.' + this.namespace, {
-                native: {
-                    devices: devices
-                }
-            });
-        }
-
-        await this.setStateChangedAsync('info.connection', !haveActiveDevices, true); //if no active device -> make green.
     }
 
     /**
