+++ src/main.d.ts
@@ -3,4 +3,3 @@ import { Device } from './lib/Device';
 export interface Mydlink extends ioBroker.Adapter {
     devices: Array<Device>;
 }
-
